+++ qiskit/providers/builtinsimulators/simulatorsjob.py
@@ -12,7 +12,7 @@ import logging
 import sys
 import functools
 
-from qiskit.backends import BaseJob, JobStatus, JobError
+from qiskit.providers import BaseJob, JobStatus, JobError
 from qiskit.qobj import validate_qobj_against_schema
 
 logger = logging.getLogger(__name__)
@@ -124,3 +124,11 @@ class SimulatorsJob(BaseJob):
     def backend(self):
         """Return the instance of the backend used for this job."""
         return self._backend
+
+    def qobj(self):
+        """Return the Qobj submitted for this job.
+
+        Returns:
+            Qobj: the Qobj submitted for this job.
+        """
+        return self._qobj