+++ qiskit/transpiler/passes/__init__.py
@@ -15,5 +15,6 @@ from .mapping.check_map import CheckMap
 from .mapping.basic_mapper import BasicMapper
 from .mapping.stochastic_mapper import StochasticMapper
 from .mapping.direction_mapper import DirectionMapper
+from .mapping.cx_direction import CXDirection
 from .mapping.unroller import Unroller
 from .mapping.lookahead_mapper import LookaheadMapper