+++ qiskit/transpiler/passes/mapping/cx_direction.py
@@ -6,8 +6,8 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """
-A direction mapper rearrenges the direction of the cx nodes to make the circuit
-compatible to the directed coupling map.
+The CX direction rearrenges the direction of the cx nodes to make the circuit
+compatible with the coupling_map.
 """
 
 from qiskit.transpiler._basepasses import TransformationPass
@@ -17,7 +17,7 @@ from qiskit.mapper import Layout
 from qiskit.extensions.standard import HGate
 
 
-class DirectionMapper(TransformationPass):
+class CXDirection(TransformationPass):
     """
      Rearranges the direction of the cx nodes to make the circuit
      compatible with the directed coupling map.
@@ -31,7 +31,7 @@ class DirectionMapper(TransformationPass):
     def __init__(self, coupling_map, initial_layout=None):
         """
         Args:
-            coupling_map (Coupling): Directed graph represented a coupling map.
+            coupling_map (CouplingMap): Directed graph represented a coupling map.
             initial_layout (Layout): The initial layout of the DAG.
         """