+++ test/python/transpiler/test_direction_selection.py
@@ -5,20 +5,20 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Test the Direction Mapper pass"""
+"""Test the CX Direction  pass"""
 
 import unittest
 
 from qiskit import ClassicalRegister, QuantumRegister, QuantumCircuit
 from qiskit.transpiler import MapperError
-from qiskit.transpiler.passes import DirectionMapper
-from qiskit.mapper import Coupling
+from qiskit.mapper import CouplingMap
+from qiskit.transpiler.passes import CXDirection
 from qiskit.converters import circuit_to_dag
 from ..common import QiskitTestCase
 
 
-class TestDirectionMapper(QiskitTestCase):
-    """ Tests the DirectionMapper pass."""
+class TestCXDirection(QiskitTestCase):
+    """ Tests the CXDirection pass."""
 
     def test_no_cnots(self):
         """ Trivial map in a circuit without entanglement
@@ -28,15 +28,15 @@ class TestDirectionMapper(QiskitTestCase):
 
          qr2:---[H]---
 
-         Coupling map: None
+         CouplingMap map: None
         """
         qr = QuantumRegister(3, 'qr')
         circuit = QuantumCircuit(qr)
         circuit.h(qr)
-        coupling = Coupling()
+        coupling = CouplingMap()
         dag = circuit_to_dag(circuit)
 
-        pass_ = DirectionMapper(coupling)
+        pass_ = CXDirection(coupling)
         after = pass_.run(dag)
 
         self.assertEqual(dag, after)
@@ -49,15 +49,15 @@ class TestDirectionMapper(QiskitTestCase):
                  |
          qr2:----.----
 
-         Coupling map: [2] <- [0] -> [1]
+         CouplingMap map: [2] <- [0] -> [1]
         """
         qr = QuantumRegister(3, 'qr')
         circuit = QuantumCircuit(qr)
         circuit.cx(qr[1], qr[2])
-        coupling = Coupling({0: [2, 1]})
+        coupling = CouplingMap([(0, 1), (0, 2)])
         dag = circuit_to_dag(circuit)
 
-        pass_ = DirectionMapper(coupling)
+        pass_ = CXDirection(coupling)
 
         with self.assertRaises(MapperError):
             pass_.run(dag)
@@ -68,15 +68,15 @@ class TestDirectionMapper(QiskitTestCase):
                  |
          qr1:----.----
 
-         Coupling map: [0] -> [1]
+         CouplingMap map: [0] -> [1]
         """
         qr = QuantumRegister(2, 'qr')
         circuit = QuantumCircuit(qr)
         circuit.cx(qr[0], qr[1])
-        coupling = Coupling({0: [1]})
+        coupling = CouplingMap([(0, 1)])
         dag = circuit_to_dag(circuit)
 
-        pass_ = DirectionMapper(coupling)
+        pass_ = CXDirection(coupling)
         after = pass_.run(dag)
 
         self.assertEqual(dag, after)
@@ -87,7 +87,7 @@ class TestDirectionMapper(QiskitTestCase):
                  |
          qr1:---(+)---
 
-         Coupling map: [0] -> [1]
+         CouplingMap map: [0] -> [1]
 
          qr0:-[H]-(+)-[H]--
                    |
@@ -96,7 +96,7 @@ class TestDirectionMapper(QiskitTestCase):
         qr = QuantumRegister(2, 'qr')
         circuit = QuantumCircuit(qr)
         circuit.cx(qr[1], qr[0])
-        coupling = Coupling({0: [1]})
+        coupling = CouplingMap([(0, 1)])
         dag = circuit_to_dag(circuit)
 
         expected = QuantumCircuit(qr)
@@ -106,7 +106,7 @@ class TestDirectionMapper(QiskitTestCase):
         expected.h(qr[0])
         expected.h(qr[1])
 
-        pass_ = DirectionMapper(coupling)
+        pass_ = CXDirection(coupling)
         after = pass_.run(dag)
 
         self.assertEqual(circuit_to_dag(expected), after)
@@ -119,7 +119,7 @@ class TestDirectionMapper(QiskitTestCase):
                     |
          cr0: ------.--
 
-         Coupling map: [0] -> [1]
+         CouplingMap map: [0] -> [1]
 
          qr0: -[H]--.--[H]-[m]-
                     |       |
@@ -133,7 +133,7 @@ class TestDirectionMapper(QiskitTestCase):
         circuit = QuantumCircuit(qr, cr)
         circuit.cx(qr[1], qr[0])
         circuit.measure(qr[0], cr[0])
-        coupling = Coupling({0: [1]})
+        coupling = CouplingMap([(0, 1)])
         dag = circuit_to_dag(circuit)
 
         expected = QuantumCircuit(qr, cr)
@@ -144,7 +144,7 @@ class TestDirectionMapper(QiskitTestCase):
         expected.h(qr[1])
         expected.measure(qr[0], cr[0])
 
-        pass_ = DirectionMapper(coupling)
+        pass_ = CXDirection(coupling)
         after = pass_.run(dag)
 
         self.assertEqual(circuit_to_dag(expected), after)